fix: refactorgin queue configuration
