+++ src/main/kotlin/it/pagopa/ecommerce/scheduler/queues/TransactionAuthRequestedEventsConsumer.kt
@@ -6,7 +6,6 @@ import com.azure.spring.messaging.storage.queue.core.StorageQueueTemplate
 import com.azure.spring.integration.storage.queue.inbound.StorageQueueMessageSource
 import org.slf4j.Logger
 import org.slf4j.LoggerFactory
-import org.springframework.beans.factory.annotation.Value
 import org.springframework.integration.annotation.ServiceActivator
 import org.springframework.integration.annotation.InboundChannelAdapter
 import org.springframework.integration.annotation.Poller
@@ -20,21 +19,6 @@ import org.springframework.context.annotation.Bean
 class TransactionAuthRequestedEventsConsumer() {
 
     var logger: Logger = LoggerFactory.getLogger(TransactionAuthRequestedEventsConsumer::class.java)
-
-  
-    @Value("\${azurestorage.queues.transactionauthrequestedevents.name}")
-    private val queueName: String? = null
-
-    @Bean
-    fun input() : DirectChannel {
-        return DirectChannel()
-    }
-
-    @Bean
-    @InboundChannelAdapter(channel = "transactionauthrequestedchannel", poller = [Poller(fixedDelay = "1000")])
-    fun storageQueueMessageSource(storageQueueTemplate: StorageQueueTemplate): StorageQueueMessageSource {
-        return StorageQueueMessageSource(queueName, storageQueueTemplate)
-    }
     
     @ServiceActivator(inputChannel = "transactionauthrequestedchannel")
     fun messageReceiver(@Payload payload: ByteArray, @Header(AzureHeaders.CHECKPOINTER) checkpointer: Checkpointer) {
